inria_project/requirements.txt
# INRIA Building Segmentation Project Requirements

# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0

# Data processing
albumentations>=1.3.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Metrics and visualization
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Segmentation models (для быстрого прототипирования)
segmentation-models-pytorch>=0.3.3

# Model training utilities
tensorboard>=2.13.0
wandb>=0.15.0  # опционально для логирования

# Utilities
pyyaml>=6.0
pandas>=2.0.0

# Web interface (для демо)
streamlit>=1.25.0
gradio>=3.50.0

# Geospatial (для работы с TIFF)
rasterio>=1.3.0  # опционально
gdal>=3.7.0  # опционально, если нужна полная поддержка GeoTIFF
